test: move statusMessage wire-byte tests to their own file
node-http.test.ts has an unrelated pre-existing failure in this container
(the issue#4295 proxy test) which masks the fail-before/pass-after signal.
The focused file matches the existing per-topic test pattern in
test/js/node/http/ and carries only the reason-phrase assertions.
